changed packages to com storngkey and added new replication binary files
