+++ server/authenticationBeans/src/main/java/com/strongkey/auth/txbeans/authorizeLdapUserBeanLocal.java
@@ -56,7 +56,7 @@ public interface authorizeLdapUserBeanLocal {
      * user - either ENC or DEC for encryption and decryption respectively
      * @return boolean value indicating either True (for authenticated) or False
      * (for unauthenticated or failure in processing)
-     * @throws com.strongauth.skce.utilities.SKCEException
+     * @throws com.strongkey.skce.utilities.SKCEException
      */
     boolean execute(Long did,
             String username,
