+++ server/fido2mds/pom.xml
@@ -3,7 +3,7 @@
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <artifactId>fidoserver</artifactId>
-        <groupId>com.strongauth</groupId>
+        <groupId>com.strongkey</groupId>
         <version>4.0</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
@@ -23,13 +23,13 @@
 
     <dependencies>
         <dependency>
-            <groupId>com.strongauth</groupId>
+            <groupId>com.strongkey</groupId>
             <artifactId>common</artifactId>
             <version>4.0</version>
             <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.strongauth</groupId>
+            <groupId>com.strongkey</groupId>
             <artifactId>crypto</artifactId>
             <version>4.0</version>
         </dependency>
