+++ server/fidoserverbeans/src/main/java/com/strongkey/skfs/messaging/persistReplicationObjectBean.java
@@ -10,9 +10,9 @@ import com.strongkey.appliance.entitybeans.Servers;
 import com.strongkey.appliance.utilities.applianceCommon;
 import com.strongkey.appliance.utilities.applianceConstants;
 import com.strongkey.appliance.utilities.strongkeyLogger;
-import com.strongauth.replication.entitybeans.Replication;
-import com.strongauth.replication.entitybeans.ReplicationPK;
-import com.strongauth.replication.txbeans.getReplicationIDLocal;
+import com.strongkey.replication.entitybeans.Replication;
+import com.strongkey.replication.entitybeans.ReplicationPK;
+import com.strongkey.replication.txbeans.getReplicationIDLocal;
 import com.strongkey.skce.utilities.skceConstants;
 import com.strongkey.skfs.txbeans.getServerBeanLocal;
 import com.strongkey.skfs.utilities.SKFEException;
