+++ server/skfe/src/main/webapp/WEB-INF/web.xml
@@ -5,21 +5,5 @@
             30
         </session-timeout>
     </session-config>
-    <!--    <servlet>  
-    <servlet-name>Jersey REST Service</servlet-name>  
-    <servlet-class>com.sun.jersey.spi.container.servlet.ServletContainer</servlet-class>  
-    <init-param>  
-      <param-name>com.sun.jersey.config.property.packages</param-name>  
-      <param-value>com.strongauth.skfews.u2f.rest</param-value>  
-    </init-param>  
-    <init-param>
-      <param-name>com.sun.jersey.spi.container.ContainerResponseFilters</param-name>
-      <param-value>com.strongauth.skfews.accesscontrol.CrossOriginResourceSharingFilter</param-value>        
-    </init-param>
-    <load-on-startup>1</load-on-startup>  
-  </servlet>  
-  <servlet-mapping>  
-    <servlet-name>Jersey REST Service</servlet-name>  
-    <url-pattern>/*</url-pattern>  
-  </servlet-mapping>-->
+
 </web-app>
